+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- njs-acme now supports ACME external account binding (EAB).
+
 ## 1.0.0 (March 25, 2024)
 
 Initial release of njs-acme.